requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
azure-storage-blob==12.23.1
python-dotenv==1.0.1
httpx==0.27.2
pypdf==4.3.1
python-multipart==0.0.12
pytesseract==0.3.13
Pillow==10.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
azure-storage-blob==12.23.1
python-dotenv==1.0.1